requirements.txt
python-telegram-bot==20.8
google-generativeai==0.5.2
flask==2.3.3
